9879 Move code to floating title bar. Move title to common folder.
