+++ src-built-in/components/floatingTitlebar/src/components/TabRegion.jsx
@@ -6,17 +6,14 @@ import Tab from "./tab";
 import { FinsembleDnDContext, FinsembleDroppable } from '@chartiq/finsemble-react-controls';
 import { FinsembleHoverDetector } from "@chartiq/finsemble-react-controls";
 import Logo from "./logo";
-
+import Title from "../../../common/windowTitle"
 import { Store, Actions } from "../stores/tabbingStore";
 import { Store as HeaderStore, Actions as HeaderActions } from "../stores/headerStore";
 
-import { debug } from "util";
-const PLACEHOLDER_TAB = {
-    windowName: "",
-    uuid: "",
-    componentType: "placeholder-tab"
-};
 let TAB_WIDTH = 300;
+//Next two items are for calculating how large the title should be within a tab.
+const ICON_AREA = 29;
+const CLOSE_BUTTON_MARGIN = 22;
 const MINIMUM_TAB_SIZE = 100;
 
 export default class TabRegion extends React.Component {
@@ -433,6 +430,7 @@ export default class TabRegion extends React.Component {
         this.setState({ hoverState: newHoverState });
     }
     renderTabs() {
+        let titleWidth = this.state.tabWidth - ICON_AREA - CLOSE_BUTTON_MARGIN;
         return this.state.tabs.map((tab, i) => {
             return <Tab
                 onClick={() => {
@@ -453,11 +451,13 @@ export default class TabRegion extends React.Component {
                 onTabDraggedOver={this.onTabDraggedOver}
                 listenForDragOver={this.props.listenForDragOver}
                 tabWidth={this.state.tabWidth}
-                title={tab.title || tab.windowName}
+                titleWidth={titleWidth}
                 windowIdentifier={tab} />
         });
     }
     renderTitle() {
+        let titleWidth = this.state.tabWidth - ICON_AREA - CLOSE_BUTTON_MARGIN;
+
         return (<div
             draggable="true"
             onDragStart={(e) => {
@@ -469,7 +469,7 @@ export default class TabRegion extends React.Component {
             className={"fsbl-header-title cq-no-drag"}>
             <FinsembleHoverDetector edge="top" hoverAction={this.hoverAction.bind(this)} />
             <Logo windowIdentifier={Actions.getWindowIdentifier()} />
-            <div className="fsbl-tab-title">{this.props.thisWindowsTitle}</div>
+            <Title titleWidth={titleWidth} windowIdentifier={FSBL.Clients.WindowClient.getWindowIdentifier()}></Title>
         </div>);
     }
     render() {
