+++ src-built-in/components/floatingTitlebar/src/components/tab.jsx
@@ -3,8 +3,7 @@ import ReactDOM from "react-dom";
 import { FinsembleHoverDetector } from "@chartiq/finsemble-react-controls";
 import Logo from "./logo";
 import { FinsembleDraggable } from "@chartiq/finsemble-react-controls";
-const ICON_AREA = 29;
-const CLOSE_BUTTON_MARGIN = 22;
+import Title from "../../../common/windowTitle"
 /**
  * This component is pretty basic. It just takes a bunch of props and renders them.
  */
@@ -32,7 +31,6 @@ export default class Tab extends React.Component {
 	componentWillUnmount() {
 	}
 	render() {
-		let titleWidth = this.props.tabWidth - ICON_AREA - CLOSE_BUTTON_MARGIN;
 		let style = {
 			width: this.props.tabWidth
 		}
@@ -58,10 +56,7 @@ export default class Tab extends React.Component {
 				}
 				<FinsembleHoverDetector edge="top" hoverAction={this.hoverAction.bind(this)} />
 				<Logo windowIdentifier={this.props.windowIdentifier} />
-				<div className="fsbl-tab-title" style={{ width: titleWidth }}>
-					{/* @todo, figure out where we're setting the title to an empty object.... */}
-					{typeof (this.props.title) === "string" ? this.props.title : ''}
-				</div>
+				<Title titleWidth={this.props.titleWidth} windowIdentifier={this.props.windowIdentifier} />
 				<div className="fsbl-tab-close" onClick={(e) => {
 					e.preventDefault();
 					e.stopPropagation();
